kageco-280 [月間カレンダー機能の分離と統合] ・MonthlyCalendarコンポーネントを新規作成し、timecard-view.tsxから分離した ・timecard-view.tsxで分離したMonthlyCalendarコンポーネントを呼び出すように修正した ・CSSなどのスタイル定義をそれぞれ適切な箇所に配置し直した ・business/timecard/index.tsxにおいて、スタイルを追加し全体のレイアウトを調整した である
